requirements.txt
requests
python-dotenv
pandas
chromadb
sentence-transformers
pytesseract
Pillow

